Apps/reviews/requirements.txt
streamlit>=1.28.0
databricks-sql-connector>=3.0.0
pandas>=2.0.0
pyarrow>=12.0.0
databricks-sdk
